+++ components/script/dom/bindings/mod.rs
@@ -138,12 +138,11 @@ pub mod buffer_source;
 pub mod callback;
 pub mod cell;
 pub mod constant;
-pub mod constructor_hook;
+pub mod constructor;
 pub mod conversions;
 pub mod error;
 pub mod finalize;
 pub mod guard;
-pub mod htmlconstructor;
 pub mod import;
 pub mod inheritance;
 pub mod interface;
